pyproject.toml
# SPDX-FileCopyrightText: 2019-2020 Magenta ApS
# SPDX-License-Identifier: MPL-2.0
[tool.poetry]
name = "ra-verse-proxy"
version = "0.1.0"
description = "AMQP Reverse Proxy (GIR Integration Point and Engine)"
authors = ["Magenta ApS <info@magenta.dk>"]
license = "MPL-2.0"
readme = "README.md"
homepage = "https://magenta.dk/"
repository = "https://git.magenta.dk/rammearkitektur/ra-verse-proxy"
keywords = ["os2mo", "amqp", "gir", "rpc"]

[tool.poetry.dependencies]
python = "^3.10"
celery = "^5.2.6"
uvicorn = "^0.17.6"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
strawberry-graphql = "^0.113.0"
httpx = "^0.23.0"
asgiref = "^3.5.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
reorder-python-imports = "^3.1.0"
mypy = "^0.960"
pre-commit = "^2.19.0"
pylint-pydantic = "^0.1.4"
watchdog = "^2.1.8"
celery = {extras = ["pytest"], version = "^5.2.6"}
httpx = "^0.23.0"
requests = "^2.27.1"
respx = "^0.19.2"
more-itertools = "^8.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopub]
git-username = "MagentaBot"
git-email = "info@magenta.dk"

[tool.pytest.ini_options]
asyncio_mode="auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = "True"
warn_unused_configs = "True"
check_untyped_defs = "True"
disallow_untyped_defs = "True"
plugins = "pydantic.mypy"

[tool.flake8]
max-line-length = 88
exclude = [
".git", "__pycache__", ".mypy_cache", ".pytest_cache", ".venv", ".gitlab"
]
